Refactored all the mockdb populators into one file, and changed the client_id to clientId
